requirements.txt
backcall==0.2.0
cloudpickle==2.1.0
cmake==3.22.5
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
igrf==13.0.2
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==7.34.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.0
six==1.16.0
spyder-kernels==1.9.4
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
wcwidth==0.2.5
xarray==0.20.2
zipp==3.8.1
